user_service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.0.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
validators==0.22.0
pytest==7.4.3
httpx==0.25.1
